+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Fixed an issue where terminal prompt validation would not trigger.
 - Fixed an issue with remote cache hydration where multiple files with the same blob hash would fail
   to write them all.
+- Fixed tag-based task dependencies not creating implicit project dependencies.
 - Fixed some task/command argument quoting issues.
 
 #### 🧩 Plugins